requirements.txt
pyTelegramBotAPI
python-dotenv
colorama
